+++ libs/util/dsg-utils/src/build-logger/build-logger.spec.ts
@@ -1,5 +1,5 @@
-import { logger as applicationLogger } from "./logging";
-import { httpClient } from "./utils/http-client";
+import { logger as applicationLogger } from "../logger/logging";
+import { httpClient } from "../http-client/http-client";
 import { LogLevel } from "@amplication/util/logging";
 import { BuildLogger } from "./build-logger";
 jest.mock("./logging");
