+++ libs/util/dsg-utils/src/build-logger/build-logger.ts
@@ -1,7 +1,7 @@
 import { BuildLogger as IBuildLogger } from "@amplication/code-gen-types";
 import { LogEntry, LogLevel } from "@amplication/util-logging";
-import { logger as applicationLogger } from "./logging";
-import { httpClient } from "./utils/http-client";
+import { logger as applicationLogger } from "../logger/logging";
+import { httpClient } from "../http-client/http-client";
 
 export class BuildLogger implements IBuildLogger {
   private async createLog(log: LogEntry): Promise<void> {
