+++ libs/util/dsg-utils/src/http-client/http-client.ts
@@ -1,7 +1,7 @@
 import http from "http";
 import https from "https";
 import axios from "axios";
-import { logger } from "../logging";
+import { logger } from "../logger/logging";
 
 const httpAgent = new http.Agent({ keepAlive: true });
 const httpsAgent = new https.Agent({ keepAlive: true });
