+++ libs/util/dsg-utils/src/index.ts
@@ -2,3 +2,6 @@ export * from "./lib/entity-util";
 export * from "./lib/dto-util";
 
 export * from "./dynamic-installation/DynamicPackageInstallationManager";
+export * from "./build-logger/build-logger";
+export * from "./logger/logging";
+export * from "./build-manager-notifier/notify-build-manager";
