+++ packages/data-service-generator/src/server/resource/dto/create-dto-module.ts
@@ -51,7 +51,7 @@ import {
 import { SORT_ORDER_ID, SORT_ORDER_MODULE } from "./sort-order.util";
 import { GRAPHQL_BIGINT_VALUE, INPUT_JSON_VALUE_KEY } from "./constants";
 import DsgContext from "../../../dsg-context";
-import { logger } from "../../../logging";
+import { logger } from "@amplication/dsg-utils";
 import { DECIMAL_JS_MODULE, DECIMAL_VALUE_ID } from "./decimal-js";
 
 const FILTERS_IMPORTABLE_NAMES = Object.fromEntries(
